backend/requirement.txt
numpy
scipy
scikit-learn==1.1.3
matplotlib
pandas
fastapi
uvicorn
twilio
openai
google-generativeai==0.3.1